normalize int(32) usage to int64 (#19)
* [add] extended godoc with 10 more examples for Model and Tensor related commands

* [add] moving from int(32) to int64 across all package

+++ examples/redisai_tls_client/redisai_tls_client.go
@@ -82,7 +82,7 @@ func main() {
 
 	// Set a tensor
 	// AI.TENSORSET foo FLOAT 2 2 VALUES 1.1 2.2 3.3 4.4
-	_ = client.TensorSet("foo", redisai.TypeFloat, []int{2, 2}, []float32{1.1, 2.2, 3.3, 4.4})
+	_ = client.TensorSet("foo", redisai.TypeFloat, []int64{2, 2}, []float32{1.1, 2.2, 3.3, 4.4})
 
 	// Get a tensor content as a slice of values
 	// dt DataType, shape []int, data interface{}, err error
